Collaborative Filtering Recommendation System
Collaborative filtering is a typical way to utilize collective intelligence and it has been widely used in many fields nowadays e.g. Amazon automatically recommends books that users might be interested in based on users’ purchase and browsing history.
The goal of the project is to learn to predict the ratings of movies on the data from the GroupLens research group : MovieLens datasets ml-100k. Three models(Memory based) are built to train the data :item-based model ,user-based model and item-user-based model. NearestNeighbors is applied to find k most similar users or movies of the target user or the movies this user has watched.